Renames gulp/s3 to gulp/cdn
