+++ gulpfile.babel.js
@@ -14,7 +14,7 @@ import bump from './gulp/bump';
 import * as clean from './gulp/clean';
 import * as rsx from './gulp/rsx';
 import * as revision from './gulp/revision';
-import * as uploadToS3 from './gulp/s3';
+import * as cdn from './gulp/cdn';
 import * as git from './gulp/git';
 import * as docker from './gulp/docker';
 import * as localization from './gulp/localization';
@@ -59,11 +59,11 @@ gulp.task('rsx:build_urls', rsx.buildUrls);
 gulp.task('rsx:codex_urls', rsx.codexUrls);
 gulp.task('revision:generate', revision.generate);
 gulp.task('revision:replace', revision.replace);
-gulp.task('cdn:upload:main', uploadToS3.main);
-gulp.task('cdn:upload:audio', uploadToS3.audio);
-gulp.task('cdn:upload:main:versioned', () => uploadToS3.main(version));
-gulp.task('cdn:upload:audio:versioned', () => uploadToS3.audio(version));
-gulp.task('cdn:upload:staging', uploadToS3.staging);
+gulp.task('cdn:upload:main', cdn.main);
+gulp.task('cdn:upload:audio', cdn.audio);
+gulp.task('cdn:upload:main:versioned', () => cdn.main(version));
+gulp.task('cdn:upload:audio:versioned', () => cdn.audio(version));
+gulp.task('cdn:upload:staging', cdn.staging);
 gulp.task('changelog', git.changelog);
 gulp.task('docker:build', docker.build);
 gulp.task('docker:tag', docker.tag);
